backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.5
python-socketio==5.10.0
psycopg2-binary==2.9.9
cryptography==41.0.7
requests==2.31.0
gunicorn==21.2.0
eventlet==0.33.3
python-dotenv==1.0.0
